Added & Integrated Wiring Class
Changed the little code we have to use values from the Wiring class,
instead of magic numbers.

PLEASE Update it with constants that you use throughout the code. It is
a lovely way to quickly and easily change values for robot tuning at
competitions when the robot has a "software problem"