I might have fixed the oscillating problem! (?)
I mostly changed small things that will be helpful for debugging. The
logic is sound, except for 1 small part:

When the robot decides to turn itself toward the target, it sets
keepTurning to true before the method was given a new angle. Instead of
setting keepTurning to true, I changed it to call turnToAngle() once.
This achieves the same thing, but will give it the right angle. You can
watch the value of keepTurning any time you want, because it is
wonderful, but you should definitely not change its value.

See you all at 5:30